refactor(chat-store): extract read-state policy
